requirements.txt
streamlit
python-dotenv
openai
google-api-python-client
streamlit
google-auth
google-auth-oauthlib
google-api-python-client
openai
python-dotenv
